task2_2 fin

+++ src/main.c
@@ -1,23 +1,28 @@
-#include "../include/io_utils.h"  // Нам нужен ввод
-#include "../include/math_ops.h"  // Нам нужна математика
-#include <stdio.h>                // Нам нужен printf
+#include "../include/io_utils.h"
+#include "../include/math_ops.h"
+#include <stdio.h>
+
 
 int main(void) {
     while (1) {
         printf("Vedite chislo 1:\n");
-        int x = readInt(); // Как работает readInt? main.c не знает! Это скрыто в io_utils.c
+        double x = readDouble();
 
         printf("Vedite chislo 2:\n");
-        int y = readInt();
+        double y = readDouble();
 
         printf("Vibirite oper (+, -, *, /). Exit - q:\n");
         char z = readChar();
 
         switch (z) {
-            case '+': sum(x, y); break;
-            case '-': sub(x, y); break;
-            case '*': mul(x, y); break;
-            case '/': divv(x, y); break;
+            case '+': sum(x, y);
+                break;
+            case '-': sub(x, y);
+                break;
+            case '*': mul(x, y);
+                break;
+            case '/': divv(x, y);
+                break;
             case 'q':
                 printf("Goodbye!\n");
                 return 0;
