+++ src/math_ops.c
@@ -1,26 +1,26 @@
 #include "../include/math_ops.h"
 #include <stdio.h>
 
-void sum(int a, int b) {
-    int result = a + b;
-    printf("Sum: %d\n", result);
+void sum(double a, double b) {
+    double result = a + b;
+    printf("Sum: %lf\n", result);
 }
 
-void sub(int a, int b) {
-    int result = a - b;
-    printf("Subtract: %d\n", result);
+void sub(double a, double b) {
+    double result = a - b;
+    printf("Subtract: %lf\n", result);
 }
 
-void mul(int a, int b) {
-    int result = a * b;
-    printf("Multiply: %d\n", result);
+void mul(double a, double b) {
+    double result = a * b;
+    printf("Multiply: %.5lf\n", result);
 }
 
-void divv(int a, int b) {
-    if (b == 0) {
+void divv(double a, double b) {
+    if (b == 0.0) {
         printf("Nylzua del na nol\n");
         return;
     }
-    int result = a / b;
-    printf("Division: %d\n", result);
+    double result = a / b;
+    printf("Division: %.5lf\n", result);
 }
